Extract code related to --prompt flag to flag_prompt.go
